made required changes
